+++ MarketPlace/Aipedi/Aipedi.csproj
@@ -64,7 +64,7 @@
     <Compile Include="Enum\OrderStatus.cs" />
     <Compile Include="Enum\OrderType.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Service\PedreiroDigitalService.cs" />
+    <Compile Include="Service\AipediService.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\MarketPlace\MarketPlace.csproj">
