+++ MarketPlace/Aipedi/Domain/deliveryAddress.cs
@@ -4,18 +4,26 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PedreiroDigital.Domain
+namespace Aipedi.Domain
 {
-    public class address
+    public class deliveryAddress
     {
         public string formattedAddress { get; set; }
         public string country { get; set; }
         public string state { get; set; }
         public string city { get; set; }
+        public coordinates coordinates { get; set; }
         public string neighborhood { get; set; }
         public string streetName { get; set; }
         public string streetNumber { get; set; }
         public string postalCode { get; set; }
+        public string reference { get; set; }
+        public string complement { get; set; }
+    }
 
+    public class coordinates
+    {
+        public string latitude { get; set; }
+        public string longitude { get; set; }
     }
 }
