+++ MarketPlace/Aipedi/Properties/AssemblyInfo.cs
